+++ src/main/java/com/hcl/appscan/cli/handlers/InvokeDynamicScan.java
@@ -29,7 +29,6 @@ import com.hcl.appscan.cli.scanners.Scanner;
 import com.hcl.appscan.cli.scanners.ValidationUtil;
 import com.hcl.appscan.sdk.CoreConstants;
 import com.hcl.appscan.sdk.logging.IProgress;
-import com.hcl.appscan.sdk.logging.Message;
 import com.hcl.appscan.sdk.presence.CloudPresenceProvider;
 import com.hcl.appscan.sdk.results.CloudResultsProvider;
 import com.hcl.appscan.sdk.results.IResultsProvider;
@@ -62,7 +61,7 @@ enum Optimization { Fast, Faster, Fastest, NoOptimization }
 enum ReportFormat { html, pdf, csv, xml }
 enum LoginType { None , Automatic , Manual }
 @Command(name = "invokedynamicscan", sortOptions = false, mixinStandardHelpOptions = true, version = "1.0",
-        description = "This command serves the purpose of configuring and triggering the initiation of a Dynamic Security Analysis Scan on AppScan on Cloud. This operation is designed to seamlessly retrieve the outcomes of the scan once it has concluded. The yielded results encompass a list of pinpointed vulnerabilities, comprehensive analytical documents, and associated URLs for these reports. Moreover, the Command Line Interface (CLI) can be customized by employing specific command line options to establish criteria for failure instances. Consequently, this enables the CLI to transmit a signal indicating success or failure to the designated pipeline in a well-defined manner." ,
+        description = "This command serves the purpose of configuring and triggering the initiation of a Dynamic Security Analysis Scan on AppScan on Cloud or AppScan 360°. This operation is designed to seamlessly retrieve the outcomes of the scan once it has concluded. The yielded results encompass a list of pinpointed vulnerabilities, comprehensive analytical documents, and associated URLs for these reports. Moreover, the Command Line Interface (CLI) can be customized by employing specific command line options to establish criteria for failure instances. Consequently, this enables the CLI to transmit a signal indicating success or failure to the designated pipeline in a well-defined manner." ,
         optionListHeading = "%n@|bold,underline Options|@:%n" , descriptionHeading = "%n@|bold,underline Description|@:%n%n",
         subcommands = {HelpCommand.class})
 public class InvokeDynamicScan implements Callable<Integer> {
@@ -72,11 +71,11 @@ public class InvokeDynamicScan implements Callable<Integer> {
     @Spec
     Model.CommandSpec spec;
 
-    @Option(names = {"--key"}, description = "[Required] AppScan on Cloud API Key", required = true , order = 1)
+    @Option(names = {"--key"}, description = "[Required] AppScan on Cloud or AppScan 360° API Key", required = true , order = 1)
     private String key;
-    @Option(names = {"--secret"}, description = "[Required] AppScan on Cloud API Secret", required = true , order = 2)
+    @Option(names = {"--secret"}, description = "[Required] AppScan on Cloud or AppScan 360° API Secret", required = true , order = 2)
     private String secret;
-    @Option(names = {"--appId"}, description = "[Required] The HCL AppScan on Cloud application that this scan will be associated with", required = true , order = 3)
+    @Option(names = {"--appId"}, description = "[Required] The HCL AppScan on Cloud or AppScan 360° application that this scan will be associated with", required = true , order = 3)
     private String appId;
     @Option(names = {"--scanName"}, description = "[Required] Specify a name to use for the scan. This value is used to distinguish this scan and its results from others.", required = true , order = 4)
     private String scanName;
@@ -90,7 +89,6 @@ public class InvokeDynamicScan implements Callable<Integer> {
     @Option(names = {"--reportFormat"},defaultValue = "html",  description = "[Optional] Specify the format for the scan result report. Valid values : ${COMPLETION-CANDIDATES}.", required = false , showDefaultValue = Visibility.ALWAYS , order = 9)
     private ReportFormat reportFormat;
     private Boolean allowIntervention;
-    @Option(names = {"--presenceId"}, description = "[Optional] For sites not available on the internet, provide the ID of the AppScan Presence that can be used for the scan.", required = false ,order = 11)
     private String presenceId;
     private Boolean waitForResults;
     private Boolean failBuildNonCompliance;
@@ -103,6 +101,38 @@ public class InvokeDynamicScan implements Callable<Integer> {
     @Option(names = {"--loginPassword"}, description = "[Optional] If your app requires login, enter valid user credentials so that Application Security on Cloud can log in to the site.", required = false , order = 17)
     private String loginPassword;
     private File loginSequenceFile;
+
+    @Option(names = {"--serviceUrl"}, description = "[Optional] AppScan Service URL", required = false , order = 19)
+    private String serviceUrl;
+
+    private Boolean acceptssl;
+
+    @Option(names = {"--acceptssl"},defaultValue = "false",  paramLabel = "BOOLEAN" , description = "[Optional] Ignore untrusted certificates when connecting to AppScan 360°. Only intended for testing purposes. Not applicable to AppScan on Cloud.", required = false ,showDefaultValue = Help.Visibility.ALWAYS , order = 20)
+    public void setAcceptssl(String value) {
+        if(null!=key && !key.startsWith("local_") && (!value.isBlank()&&!"false".equalsIgnoreCase(value))){
+            logger.warn(messageBundle.getString("error.acceptssl.without.a360"));
+        }
+        if(null!=key && key.startsWith("local_")){
+            boolean invalid = !"true".equalsIgnoreCase(value) && !"false".equalsIgnoreCase(value);
+
+            if (invalid) {
+                throw new ParameterException(spec.commandLine(),
+                        String.format(messageBundle.getString("error.invalid.acceptssl"), value));
+            }else if ("true".equalsIgnoreCase(value) && serviceUrl==null){
+                throw new ParameterException(spec.commandLine(),
+                        String.format(messageBundle.getString("error.acceptssl.without.serviceurl")));
+            }
+            acceptssl = Boolean.parseBoolean(value);
+        }
+    }
+    @Option(names = {"--presenceId"}, description = "[Optional] For sites not available on the internet, provide the ID of the AppScan Presence that can be used for the scan. This option is applicable for AppScan on CLoud only.", required = false ,order = 11)
+    public void setPresenceId(String value) {
+        if(key.startsWith("local_")){
+            throw new ParameterException(spec.commandLine(),
+                    String.format(messageBundle.getString("error.invalid.presenceId.withlocalkey"), value));
+        }
+        presenceId = value;
+    }
 
     @Option(names = {"--scanFile"},  description = "[Optional] The path to a scan template file (.scan or .scant).", required = false ,showDefaultValue = Visibility.ALWAYS , order = 14)
     public void setScanFile(String filepath) {
@@ -162,14 +192,20 @@ public class InvokeDynamicScan implements Callable<Integer> {
         }
         waitForResults = Boolean.parseBoolean(value);
     }
-    @Option(names = {"--allowIntervention"},defaultValue = "false",  paramLabel = "BOOLEAN" , description = "[Optional] When set to true, our scan enablement team will step in if the scan fails, or if no issues are found, and try to fix the configuration. This may delay the scan result.", required = false ,showDefaultValue = Visibility.ALWAYS , order = 10)
+    @Option(names = {"--allowIntervention"},defaultValue = "false",  paramLabel = "BOOLEAN" , description = "[Optional] When set to true, our scan enablement team will step in if the scan fails, or if no issues are found, and try to fix the configuration. This may delay the scan result. This option is valid only for AppScan on Cloud scans.", required = false ,showDefaultValue = Visibility.ALWAYS , order = 10)
     public void setAllowIntervention(String value) {
+        if(null!=key && key.startsWith("local_") && (!value.isBlank()&&!"false".equalsIgnoreCase(value))){
+            throw new ParameterException(spec.commandLine(),
+                    String.format(messageBundle.getString("error.invalid.allowIntervention.withlocalkey")));
+        }
+
         boolean invalid = !"true".equalsIgnoreCase(value) && !"false".equalsIgnoreCase(value);
 
         if (invalid) {
             throw new ParameterException(spec.commandLine(),
                     String.format(messageBundle.getString("error.invalid.allowIntervention"), value));
         }
+
         allowIntervention = Boolean.parseBoolean(value);
     }
     @Option(names = {"--emailNotification"}, defaultValue = "false", paramLabel = "BOOLEAN" , description = "[Optional] Send the user an email when analysis is complete. Valid values : true , false", required = false , showDefaultValue = Visibility.ALWAYS , order = 8)
@@ -281,6 +317,20 @@ public class InvokeDynamicScan implements Callable<Integer> {
     private  Optional<ScanResults> runScanAndGetResults() throws Exception {
 
         CloudAuthenticationHandler authHandler = new CloudAuthenticationHandler();
+        if(null!=serviceUrl && key.startsWith("local_")){
+            if(serviceUrl.endsWith("/")){
+                serviceUrl = serviceUrl.substring(0, serviceUrl.length()-1);
+            }
+            boolean isValidURL = ValidationUtil.checkASoCConnectivity(serviceUrl,acceptssl);
+            if(!isValidURL){
+                throw new ParameterException(spec.commandLine(),
+                        String.format(messageBundle.getString("error.unreachable.serviceurl")));
+            }
+            authHandler = new CloudAuthenticationHandler(serviceUrl , acceptssl);
+        }else{
+            authHandler = new CloudAuthenticationHandler();
+        }
+
         try {
             boolean isAuthenticated = authHandler.updateCredentials(key, secret);
 
@@ -467,8 +517,18 @@ public class InvokeDynamicScan implements Callable<Integer> {
         try{
             IScanServiceProvider scanServiceProvider = scan.getServiceProvider();
             while (m_scanStatus != null && (m_scanStatus.equalsIgnoreCase(CoreConstants.INQUEUE) || m_scanStatus.equalsIgnoreCase(CoreConstants.RUNNING) || m_scanStatus.equalsIgnoreCase(CoreConstants.UNKNOWN) || m_scanStatus.equalsIgnoreCase(CoreConstants.PAUSING) || m_scanStatus.equalsIgnoreCase(CoreConstants.PAUSED)) && requestCounter < 10) {
-                String asocServerUrl = authHandler.getServer();
-                boolean isASoCServerReachable = ValidationUtil.checkASoCConnectivity(asocServerUrl);
+                String asocServerUrl;
+                boolean isASoCServerReachable;
+
+                if(serviceUrl !=null && key.startsWith("local_")){
+                    asocServerUrl = serviceUrl;
+                    isASoCServerReachable = ValidationUtil.checkASoCConnectivity(asocServerUrl,acceptssl);
+                }else{
+                    asocServerUrl = authHandler.getServer();
+                    isASoCServerReachable = ValidationUtil.checkASoCConnectivity(asocServerUrl,false);
+                }
+
+
                 if(!isASoCServerReachable){
                     m_scanStatus = CoreConstants.UNKNOWN;
                 }else{
@@ -523,8 +583,11 @@ public class InvokeDynamicScan implements Callable<Integer> {
         } else {
 
             logger.info(messageBundle.getString("info.scan.completed"), m_scanStatus);
-
-            String asocAppUrl = authHandler.getServer() + "main/myapps/" + appId + "/scans/" + scan.getScanId();
+            String scanUrlPrefix = "";
+            if(!authHandler.getServer().endsWith("/")){
+                scanUrlPrefix= authHandler.getServer()+"/";
+            }
+            String asocAppUrl = scanUrlPrefix + "main/myapps/" + appId + "/scans/" + scan.getScanId();
             ScanResults scanresults = new ScanResults(provider, scan.getName(), provider.getStatus(), provider.getFindingsCount(), provider.getCriticalCount(), provider.getHighCount(), provider.getMediumCount(), provider.getLowCount(), provider.getInfoCount(), asocAppUrl);
             results = Optional.of(scanresults);