+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Release Summary
 =========
 1.2.0 (December, 2024)
 ------
-* Added support for AppScan 360
+* Added support for HCL AppScan 360°
 * Bug fixes
 
 1.1.0 (March, 2024)