+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Release Summary
 =========
 1.2.0 (December, 2024)
 ------
-* Added support for HCL AppScan 360°
+* Support for executing DAST Scans via HCL AppScan 360° v1.3 and above.
 * Bug fixes
 
 1.1.0 (March, 2024)