+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Release Summary
 =========
+1.2.0 (December, 2024)
+------
+* Support for executing DAST Scans via HCL AppScan 360° v1.3 and above.
+* Bug fixes
+
 1.1.0 (March, 2024)
 ------
 * Migration to version 4 of ASoC REST APIs